removed unused MessageQueue from TimerComponent fixed Java2dRenderer to rotate before scale to proeper transformation. fixed PhysicsRigidBodyEntityBuilder to cast to int the CollisionFilterGroups short constants.
git-svn-id: https://svn.gemserk.com/componentsengine/trunk/componentsengine@235 69e4d4fc-dda9-4998-a343-e80dfd0be058

+++ jbulletmodule/src/main/java/com/gemserk/componentsengine/jbullet/entities/PhysicsRigidBodyEntityBuilder.java
@@ -30,8 +30,8 @@ public class PhysicsRigidBodyEntityBuilder extends EntityBuilder {
 
 		tags("physics", "rigidBody");
 
-		property("collisionFilterGroup", parameters.get("collisionFilterGroup") != null ? parameters.get("collisionFilterGroup") : CollisionFilterGroups.DEFAULT_FILTER);
-		property("collisionFilterMask", parameters.get("collisionFilterMask") != null ? parameters.get("collisionFilterMask") : CollisionFilterGroups.ALL_FILTER);
+		property("collisionFilterGroup", parameters.get("collisionFilterGroup") != null ? parameters.get("collisionFilterGroup") : (int)CollisionFilterGroups.DEFAULT_FILTER);
+		property("collisionFilterMask", parameters.get("collisionFilterMask") != null ? parameters.get("collisionFilterMask") : (int)CollisionFilterGroups.ALL_FILTER);
 
 		final String prefix = (String) (parameters.get("prefix") != null ? parameters.get("prefix") : "physics");
